backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
